requirements.txt
ImageHash
Pillow
requests
numpy
scikit-image
faiss-cpu
streamlit